admin/requirements.txt
flask==3.0.0
flask-login==0.6.3
flask-bcrypt==1.0.1
flask-wtf==1.2.1
flask-limiter==3.5.0
flask-session==0.5.0
python-dotenv==1.0.0
bleach==6.1.0
email-validator==2.1.0
redis==5.0.1
